refactor: rename board to home
