+++ corecomponent/androidcomponent/src/main/res/menu/menu_nav_bottom.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/board"
+        android:id="@+id/home"
         android:icon="@drawable/ic_home"
         android:title="@string/title_board"/>
