+++ features/home/src/androidTest/java/com/techcafe/todone/home/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.techcafe.todone.board
+package com.techcafe.todone.home
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
