+++ features/home/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.techcafe.todone.board" />
+    package="com.techcafe.todone.home" />
