+++ features/home/src/main/java/com/techcafe/todone/home/model/SampleModel.kt
@@ -1,4 +1,4 @@
-package com.techcafe.todone.board.model
+package com.techcafe.todone.home.model
 
 data class SampleModel(
     val title: String?,
