+++ features/home/src/main/java/com/techcafe/todone/home/ui/BoardViewHolder.kt
@@ -1,8 +1,8 @@
-package com.techcafe.todone.board.ui
+package com.techcafe.todone.home.ui
 
 import androidx.recyclerview.widget.RecyclerView
-import com.techcafe.todone.board.databinding.ViewItemBinding
-import com.techcafe.todone.board.model.SampleModel
+import com.techcafe.todone.home.databinding.ViewItemBinding
+import com.techcafe.todone.home.model.SampleModel
 
 internal class BoardViewHolder(private val binding: ViewItemBinding) :
     RecyclerView.ViewHolder(binding.root) {
