+++ features/home/src/main/java/com/techcafe/todone/home/viewModel/BoardCreateViewModel.kt
@@ -1,4 +1,4 @@
-package com.techcafe.todone.board.viewModel
+package com.techcafe.todone.home.viewModel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
