+++ features/home/src/main/res/layout/fragment_create.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.techcafe.todone.board.viewModel.BoardCreateViewModel"/>
+            type="com.techcafe.todone.home.viewModel.BoardCreateViewModel"/>
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
